% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_locations.R
\name{amltr_list_locations}
\alias{amltr_list_locations}
\title{Lists information about the supported locations for this service.}
\usage{
amltr_list_locations(projectId, pageToken = NULL, pageSize = NULL,
  filter = NULL)
}
\arguments{
\item{projectId}{The resource that owns the locations collection, if applicable}

\item{pageToken}{The standard list page token}

\item{pageSize}{The standard list page size}

\item{filter}{The standard list filter}
}
\description{
Lists information about the supported locations for this service.
}
