% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{AnnotationSpec}
\alias{AnnotationSpec}
\title{AnnotationSpec Object}
\usage{
AnnotationSpec(name = NULL, displayName = NULL, exampleCount = NULL)
}
\arguments{
\item{name}{Output only}

\item{displayName}{Required}

\item{exampleCount}{Output only}
}
\value{
AnnotationSpec object
}
\description{
AnnotationSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A definition of an annotation.
}
\concept{AnnotationSpec functions}
