% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ArrayStats}
\alias{ArrayStats}
\title{ArrayStats Object}
\usage{
ArrayStats(memberStats = NULL)
}
\arguments{
\item{memberStats}{Stats of all the values of all arrays, as if they were a single long}
}
\value{
ArrayStats object
}
\description{
ArrayStats Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The data statistics of a series of ARRAY values.
}
\concept{ArrayStats functions}
