% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_functions.R
\docType{package}
\name{automl_googleAuthR}
\alias{automl_googleAuthR}
\alias{automl_googleAuthR-package}
\title{Cloud AutoML API
Train high-quality custom machine learning models with minimum effort and machine learning expertise.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2019-04-28 16:49:49
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}
}
