% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{BatchPredictOperationMetadata}
\alias{BatchPredictOperationMetadata}
\title{BatchPredictOperationMetadata Object}
\usage{
BatchPredictOperationMetadata(outputInfo = NULL, inputConfig = NULL)
}
\arguments{
\item{outputInfo}{Output only}

\item{inputConfig}{Output only}
}
\value{
BatchPredictOperationMetadata object
}
\description{
BatchPredictOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of BatchPredict operation.
}
\concept{BatchPredictOperationMetadata functions}
