% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{BatchPredictOutputInfo}
\alias{BatchPredictOutputInfo}
\title{BatchPredictOutputInfo Object}
\usage{
BatchPredictOutputInfo(bigqueryOutputDataset = NULL,
  gcsOutputDirectory = NULL)
}
\arguments{
\item{bigqueryOutputDataset}{The path of the BigQuery dataset created, in bq://projectId}

\item{gcsOutputDirectory}{The full path of the Google Cloud Storage directory created, into which}
}
\value{
BatchPredictOutputInfo object
}
\description{
BatchPredictOutputInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Further describes this batch predict's output.SupplementsBatchPredictOutputConfig.
}
\concept{BatchPredictOutputInfo functions}
