% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{BatchPredictRequest}
\alias{BatchPredictRequest}
\title{BatchPredictRequest Object}
\usage{
BatchPredictRequest(BatchPredictRequest.params = NULL, params = NULL,
  inputConfig = NULL, outputConfig = NULL)
}
\arguments{
\item{BatchPredictRequest.params}{The \link{BatchPredictRequest.params} object or list of objects}

\item{params}{Additional domain-specific parameters for the predictions, any string must}

\item{inputConfig}{Required}

\item{outputConfig}{Required}
}
\value{
BatchPredictRequest object
}
\description{
BatchPredictRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for PredictionService.BatchPredict.
}
\seealso{
Other BatchPredictRequest functions: \code{\link{BatchPredictRequest.params}},
  \code{\link{projects.locations.models.batchPredict}}
}
\concept{BatchPredictRequest functions}
