% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{BatchPredictResult}
\alias{BatchPredictResult}
\title{BatchPredictResult Object}
\usage{
BatchPredictResult()
}
\value{
BatchPredictResult object
}
\description{
BatchPredictResult Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Result of the Batch Predict. This message is returned inresponse of the operation returnedby the PredictionService.BatchPredict.
}
\concept{BatchPredictResult functions}
