% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{BigQueryDestination}
\alias{BigQueryDestination}
\title{BigQueryDestination Object}
\usage{
BigQueryDestination(outputUri = NULL)
}
\arguments{
\item{outputUri}{Required}
}
\value{
BigQueryDestination object
}
\description{
BigQueryDestination Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The BigQuery location for the output content.
}
\concept{BigQueryDestination functions}
