% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{BigQuerySource}
\alias{BigQuerySource}
\title{BigQuerySource Object}
\usage{
BigQuerySource(inputUri = NULL)
}
\arguments{
\item{inputUri}{Required}
}
\value{
BigQuerySource object
}
\description{
BigQuerySource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The BigQuery location for the input content.
}
\concept{BigQuerySource functions}
