% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{BoundingBoxMetricsEntry}
\alias{BoundingBoxMetricsEntry}
\title{BoundingBoxMetricsEntry Object}
\usage{
BoundingBoxMetricsEntry(meanAveragePrecision = NULL,
  confidenceMetricsEntries = NULL, iouThreshold = NULL)
}
\arguments{
\item{meanAveragePrecision}{Output only}

\item{confidenceMetricsEntries}{Output only}

\item{iouThreshold}{Output only}
}
\value{
BoundingBoxMetricsEntry object
}
\description{
BoundingBoxMetricsEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Bounding box matching model metrics for a single intersection-over-unionthreshold and multiple label match confidence thresholds.
}
\concept{BoundingBoxMetricsEntry functions}
