% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{BoundingBoxMetricsEntryConfidenceMetricsEntry}
\alias{BoundingBoxMetricsEntryConfidenceMetricsEntry}
\title{BoundingBoxMetricsEntryConfidenceMetricsEntry Object}
\usage{
BoundingBoxMetricsEntryConfidenceMetricsEntry(recall = NULL,
  f1Score = NULL, precision = NULL, confidenceThreshold = NULL)
}
\arguments{
\item{recall}{Output only}

\item{f1Score}{Output only}

\item{precision}{Output only}

\item{confidenceThreshold}{Output only}
}
\value{
BoundingBoxMetricsEntryConfidenceMetricsEntry object
}
\description{
BoundingBoxMetricsEntryConfidenceMetricsEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metrics for a single confidence threshold.
}
\concept{BoundingBoxMetricsEntryConfidenceMetricsEntry functions}
