% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{BoundingPoly}
\alias{BoundingPoly}
\title{BoundingPoly Object}
\usage{
BoundingPoly(normalizedVertices = NULL)
}
\arguments{
\item{normalizedVertices}{Output only}
}
\value{
BoundingPoly object
}
\description{
BoundingPoly Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A bounding polygon of a detected object on a plane.On output both vertices and normalized_vertices are provided.The polygon is formed by connecting vertices in the order they are listed.
}
\concept{BoundingPoly functions}
