% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{CategoryStats}
\alias{CategoryStats}
\title{CategoryStats Object}
\usage{
CategoryStats(topCategoryStats = NULL)
}
\arguments{
\item{topCategoryStats}{The statistics of the top 20 CATEGORY values, ordered by}
}
\value{
CategoryStats object
}
\description{
CategoryStats Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The data statistics of a series of CATEGORY values.
}
\concept{CategoryStats functions}
