% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ClassificationAnnotation}
\alias{ClassificationAnnotation}
\title{ClassificationAnnotation Object}
\usage{
ClassificationAnnotation(score = NULL)
}
\arguments{
\item{score}{Output only}
}
\value{
ClassificationAnnotation object
}
\description{
ClassificationAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains annotation details specific to classification.
}
\concept{ClassificationAnnotation functions}
