% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ClassificationEvaluationMetricsConfusionMatrixRow}
\alias{ClassificationEvaluationMetricsConfusionMatrixRow}
\title{ClassificationEvaluationMetricsConfusionMatrixRow Object}
\usage{
ClassificationEvaluationMetricsConfusionMatrixRow(exampleCount = NULL)
}
\arguments{
\item{exampleCount}{Output only}
}
\value{
ClassificationEvaluationMetricsConfusionMatrixRow object
}
\description{
ClassificationEvaluationMetricsConfusionMatrixRow Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Output only. A row in the confusion matrix.
}
\concept{ClassificationEvaluationMetricsConfusionMatrixRow functions}
