% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ColumnSpec}
\alias{ColumnSpec}
\title{ColumnSpec Object}
\usage{
ColumnSpec(dataStats = NULL, displayName = NULL, etag = NULL,
  dataType = NULL, name = NULL, topCorrelatedColumns = NULL)
}
\arguments{
\item{dataStats}{Output only}

\item{displayName}{Output only}

\item{etag}{Used to perform consistent read-modify-write updates}

\item{dataType}{The data type of elements stored in the column}

\item{name}{Output only}

\item{topCorrelatedColumns}{Output only}
}
\value{
ColumnSpec object
}
\description{
ColumnSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A representation of a column in a relational table. When listing them, column specs are returned in the same order in which they weregiven on import .Used by:  *   Tables
}
\seealso{
Other ColumnSpec functions: \code{\link{projects.locations.datasets.tableSpecs.columnSpecs.patch}}
}
\concept{ColumnSpec functions}
