% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ConfusionMatrix}
\alias{ConfusionMatrix}
\title{ConfusionMatrix Object}
\usage{
ConfusionMatrix(annotationSpecId = NULL, row = NULL)
}
\arguments{
\item{annotationSpecId}{Output only}

\item{row}{Output only}
}
\value{
ConfusionMatrix object
}
\description{
ConfusionMatrix Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Confusion matrix of the model running the classification.
}
\concept{ConfusionMatrix functions}
