% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{CorrelatedColumn}
\alias{CorrelatedColumn}
\title{CorrelatedColumn Object}
\usage{
CorrelatedColumn(correlationStats = NULL, columnSpecId = NULL)
}
\arguments{
\item{correlationStats}{Correlation between this and the in-context column}

\item{columnSpecId}{The column_spec_id of the correlated column, which belongs to the same}
}
\value{
CorrelatedColumn object
}
\description{
CorrelatedColumn Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Identifies the table's column, and its correlation with the column thisColumnSpec describes.
}
\concept{CorrelatedColumn functions}
