% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{CorrelationStats}
\alias{CorrelationStats}
\title{CorrelationStats Object}
\usage{
CorrelationStats(cramersV = NULL)
}
\arguments{
\item{cramersV}{The correlation value using the Cramer's V measure}
}
\value{
CorrelationStats object
}
\description{
CorrelationStats Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A correlation statistics between two series of DataType values. The seriesmay have differing DataType-s, but within a single series the DataType mustbe the same.
}
\concept{CorrelationStats functions}
