% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{CreateModelOperationMetadata}
\alias{CreateModelOperationMetadata}
\title{CreateModelOperationMetadata Object}
\usage{
CreateModelOperationMetadata()
}
\value{
CreateModelOperationMetadata object
}
\description{
CreateModelOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of CreateModel operation.
}
\concept{CreateModelOperationMetadata functions}
