% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset.R
\name{Dataset}
\alias{Dataset}
\title{Dataset Object}
\usage{
Dataset(description = NULL, imageClassificationDatasetMetadata = NULL,
  etag = NULL, textSentimentDatasetMetadata = NULL, name = NULL,
  imageObjectDetectionDatasetMetadata = NULL, exampleCount = NULL,
  translationDatasetMetadata = NULL, displayName = NULL,
  textClassificationDatasetMetadata = NULL, createTime = NULL,
  textExtractionDatasetMetadata = NULL,
  videoClassificationDatasetMetadata = NULL,
  tablesDatasetMetadata = NULL)
}
\arguments{
\item{description}{User-provided description of the dataset}

\item{imageClassificationDatasetMetadata}{Metadata for a dataset used for image classification}

\item{etag}{Used to perform consistent read-modify-write updates}

\item{textSentimentDatasetMetadata}{Metadata for a dataset used for text sentiment}

\item{name}{Output only}

\item{imageObjectDetectionDatasetMetadata}{Metadata for a dataset used for image object detection}

\item{exampleCount}{Output only}

\item{translationDatasetMetadata}{Metadata for a dataset used for translation}

\item{displayName}{Required}

\item{textClassificationDatasetMetadata}{Metadata for a dataset used for text classification}

\item{createTime}{Output only}

\item{textExtractionDatasetMetadata}{Metadata for a dataset used for text extraction}

\item{videoClassificationDatasetMetadata}{Metadata for a dataset used for video classification}

\item{tablesDatasetMetadata}{Metadata for a dataset used for Tables}
}
\value{
Dataset object
}
\description{
Dataset Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A workspace for solving a single, particular machine learning (ML) problem.A workspace contains examples that may be annotated.
}
\seealso{
Other Dataset functions: \code{\link{gcat_create_dataset}}
}
\concept{Dataset functions}
