% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{DataStats}
\alias{DataStats}
\title{DataStats Object}
\usage{
DataStats(stringStats = NULL, arrayStats = NULL, structStats = NULL,
  categoryStats = NULL, distinctValueCount = NULL,
  nullValueCount = NULL, float64Stats = NULL, timestampStats = NULL)
}
\arguments{
\item{stringStats}{The statistics for STRING DataType}

\item{arrayStats}{The statistics for ARRAY DataType}

\item{structStats}{The statistics for STRUCT DataType}

\item{categoryStats}{The statistics for CATEGORY DataType}

\item{distinctValueCount}{The number of distinct values}

\item{nullValueCount}{The number of values that are null}

\item{float64Stats}{The statistics for FLOAT64 DataType}

\item{timestampStats}{The statistics for TIMESTAMP DataType}
}
\value{
DataStats object
}
\description{
DataStats Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The data statistics of a series of values that share the same DataType.
}
\concept{DataStats functions}
