% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{DataType}
\alias{DataType}
\title{DataType Object}
\usage{
DataType(nullable = NULL, listElementType = NULL, typeCode = NULL,
  structType = NULL, timeFormat = NULL)
}
\arguments{
\item{nullable}{If true, this DataType can also be `NULL`}

\item{listElementType}{If type_code == ARRAY,}

\item{typeCode}{Required}

\item{structType}{If type_code == STRUCT, then `struct_type`}

\item{timeFormat}{If type_code == TIMESTAMP}
}
\value{
DataType object
}
\description{
DataType Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Indicated the type of data that can be stored in a structured data entity(e.g. a table).
}
\concept{DataType functions}
