% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{DeleteOperationMetadata}
\alias{DeleteOperationMetadata}
\title{DeleteOperationMetadata Object}
\usage{
DeleteOperationMetadata()
}
\value{
DeleteOperationMetadata object
}
\description{
DeleteOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of operations that perform deletes of any entities.
}
\concept{DeleteOperationMetadata functions}
