% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{DeployModelOperationMetadata}
\alias{DeployModelOperationMetadata}
\title{DeployModelOperationMetadata Object}
\usage{
DeployModelOperationMetadata()
}
\value{
DeployModelOperationMetadata object
}
\description{
DeployModelOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of DeployModel operation.
}
\concept{DeployModelOperationMetadata functions}
