% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{DeployModelRequest}
\alias{DeployModelRequest}
\title{DeployModelRequest Object}
\usage{
DeployModelRequest(imageObjectDetectionModelDeploymentMetadata = NULL)
}
\arguments{
\item{imageObjectDetectionModelDeploymentMetadata}{Model deployment metadata specific to Image Object Detection}
}
\value{
DeployModelRequest object
}
\description{
DeployModelRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for AutoMl.DeployModel.
}
\seealso{
Other DeployModelRequest functions: \code{\link{projects.locations.models.deploy}}
}
\concept{DeployModelRequest functions}
