% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{Document}
\alias{Document}
\title{Document Object}
\usage{
Document(inputConfig = NULL)
}
\arguments{
\item{inputConfig}{An input config specifying the content of the document}
}
\value{
Document object
}
\description{
Document Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A structured text document e.g. a PDF.
}
\concept{Document functions}
