% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{DocumentInputConfig}
\alias{DocumentInputConfig}
\title{DocumentInputConfig Object}
\usage{
DocumentInputConfig(gcsSource = NULL)
}
\arguments{
\item{gcsSource}{The Google Cloud Storage location of the document file}
}
\value{
DocumentInputConfig object
}
\description{
DocumentInputConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Input configuration of a Document.
}
\concept{DocumentInputConfig functions}
