% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{DoubleRange}
\alias{DoubleRange}
\title{DoubleRange Object}
\usage{
DoubleRange(start = NULL, end = NULL)
}
\arguments{
\item{start}{Start of the range, inclusive}

\item{end}{End of the range, exclusive}
}
\value{
DoubleRange object
}
\description{
DoubleRange Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A range between two double numbers.
}
\concept{DoubleRange functions}
