% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ExportDataOperationMetadata}
\alias{ExportDataOperationMetadata}
\title{ExportDataOperationMetadata Object}
\usage{
ExportDataOperationMetadata(outputInfo = NULL)
}
\arguments{
\item{outputInfo}{Output only}
}
\value{
ExportDataOperationMetadata object
}
\description{
ExportDataOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of ExportData operation.
}
\concept{ExportDataOperationMetadata functions}
