% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ExportDataRequest}
\alias{ExportDataRequest}
\title{ExportDataRequest Object}
\usage{
ExportDataRequest(outputConfig = NULL)
}
\arguments{
\item{outputConfig}{Required}
}
\value{
ExportDataRequest object
}
\description{
ExportDataRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for AutoMl.ExportData.
}
\seealso{
Other ExportDataRequest functions: \code{\link{projects.locations.datasets.exportData}}
}
\concept{ExportDataRequest functions}
