% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ExportEvaluatedExamplesOperationMetadata}
\alias{ExportEvaluatedExamplesOperationMetadata}
\title{ExportEvaluatedExamplesOperationMetadata Object}
\usage{
ExportEvaluatedExamplesOperationMetadata(outputInfo = NULL)
}
\arguments{
\item{outputInfo}{Output only}
}
\value{
ExportEvaluatedExamplesOperationMetadata object
}
\description{
ExportEvaluatedExamplesOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of EvaluatedExamples operation.
}
\concept{ExportEvaluatedExamplesOperationMetadata functions}
