% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ExportEvaluatedExamplesOutputInfo}
\alias{ExportEvaluatedExamplesOutputInfo}
\title{ExportEvaluatedExamplesOutputInfo Object}
\usage{
ExportEvaluatedExamplesOutputInfo(bigqueryOutputDataset = NULL)
}
\arguments{
\item{bigqueryOutputDataset}{The path of the BigQuery dataset created, in bq://projectId}
}
\value{
ExportEvaluatedExamplesOutputInfo object
}
\description{
ExportEvaluatedExamplesOutputInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Further describes the output of the evaluated examples export.SupplementsExportEvaluatedExamplesOutputConfig.
}
\concept{ExportEvaluatedExamplesOutputInfo functions}
