% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ExportEvaluatedExamplesRequest}
\alias{ExportEvaluatedExamplesRequest}
\title{ExportEvaluatedExamplesRequest Object}
\usage{
ExportEvaluatedExamplesRequest(outputConfig = NULL)
}
\arguments{
\item{outputConfig}{Required}
}
\value{
ExportEvaluatedExamplesRequest object
}
\description{
ExportEvaluatedExamplesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for AutoMl.ExportEvaluatedExamples.
}
\seealso{
Other ExportEvaluatedExamplesRequest functions: \code{\link{projects.locations.models.exportEvaluatedExamples}}
}
\concept{ExportEvaluatedExamplesRequest functions}
