% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ExportModelOperationMetadata}
\alias{ExportModelOperationMetadata}
\title{ExportModelOperationMetadata Object}
\usage{
ExportModelOperationMetadata(outputInfo = NULL)
}
\arguments{
\item{outputInfo}{Output only}
}
\value{
ExportModelOperationMetadata object
}
\description{
ExportModelOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of ExportModel operation.
}
\concept{ExportModelOperationMetadata functions}
