% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ExportModelOutputInfo}
\alias{ExportModelOutputInfo}
\title{ExportModelOutputInfo Object}
\usage{
ExportModelOutputInfo(gcsOutputDirectory = NULL)
}
\arguments{
\item{gcsOutputDirectory}{The full path of the Google Cloud Storage directory created, into which}
}
\value{
ExportModelOutputInfo object
}
\description{
ExportModelOutputInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Further describes the output of model export.SupplementsModelExportOutputConfig.
}
\concept{ExportModelOutputInfo functions}
