% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ExportModelRequest}
\alias{ExportModelRequest}
\title{ExportModelRequest Object}
\usage{
ExportModelRequest(outputConfig = NULL)
}
\arguments{
\item{outputConfig}{Required}
}
\value{
ExportModelRequest object
}
\description{
ExportModelRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for AutoMl.ExportModel.Models need to be enabled for exporting, otherwise an error code will bereturned.
}
\seealso{
Other ExportModelRequest functions: \code{\link{projects.locations.models.export}}
}
\concept{ExportModelRequest functions}
