% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{Float64Stats}
\alias{Float64Stats}
\title{Float64Stats Object}
\usage{
Float64Stats(histogramBuckets = NULL, mean = NULL,
  standardDeviation = NULL, quantiles = NULL)
}
\arguments{
\item{histogramBuckets}{Histogram buckets of the data series}

\item{mean}{The mean of the series}

\item{standardDeviation}{The standard deviation of the series}

\item{quantiles}{Ordered from 0 to k k-quantile values of the data series of n values}
}
\value{
Float64Stats object
}
\description{
Float64Stats Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The data statistics of a series of FLOAT64 values.
}
\concept{Float64Stats functions}
