% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{gcat_batch_predict}
\alias{gcat_batch_predict}
\title{Perform a batch prediction. Unlike the online Predict,
batchprediction result won't be immediately available in the response.
Instead,a long running operation object is returned.}
\usage{
gcat_batch_predict(modelDisplayName, inputSource, outputTarget)
}
\arguments{
\item{modelDisplayName}{the name of the model shown in the interface}

\item{inputSource}{Google Cloud Storage URIs to input files, up to 2000 characters long. Full object path, e.g. gs://bucket/directory/object.csv}

\item{outputTarget}{STRING Required. Google Cloud Storage URI to output directory, up to 2000 characters long. Prefix path: gs://bucket/directory The requesting user must have write permission to the bucket. The directory is created if it doesn't exist.}
}
\description{
Perform a batch prediction. Unlike the online Predict,
batchprediction result won't be immediately available in the response.
Instead,a long running operation object is returned.
}
\seealso{
Other Model functions: 
\code{\link{gcat_create_model}()},
\code{\link{gcat_get_model}()},
\code{\link{gcat_list_model_evaluations}()},
\code{\link{gcat_list_models}()}
}
\concept{Model functions}
