% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{gcat_create_dataset}
\alias{gcat_create_dataset}
\title{Creates a dataset}
\usage{
gcat_create_dataset(projectId, locationId, displayName, parent)
}
\arguments{
\item{parent}{}

\item{location}{location of GCP resources}
}
\description{
Creates a dataset
}
