% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset.R
\name{gcat_create_dataset_do_call}
\alias{gcat_create_dataset_do_call}
\title{Creates a dataset (Internal API call).}
\usage{
gcat_create_dataset_do_call(Dataset, parent)
}
\arguments{
\item{parent}{The resource name of the project to create the dataset for}
}
\description{
Creates a dataset (Internal API call).
}
\seealso{
Other Dataset functions: \code{\link{Dataset}}
}
\concept{Dataset functions}
