% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{gcat_delete_dataset}
\alias{gcat_delete_dataset}
\title{Deletes a dataset}
\usage{
gcat_delete_dataset(
  projectId = gcat_project_get(),
  locationId = gcat_region_get(),
  displayName
)
}
\arguments{
\item{projectId}{GCP project id}

\item{locationId}{location of GCP resources}

\item{displayName}{the name of the dataset that is shown in the interface.}
}
\description{
Deletes a dataset
}
