% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{gcat_get_location}
\alias{gcat_get_location}
\title{Get a location}
\usage{
gcat_get_location(projectId, locationId)
}
\arguments{
\item{projectId}{the GCP project Id containing location to fetch}

\item{locationId}{location of GCP resources}
}
\description{
Gets information about a location
}
