% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{gcat_get_model}
\alias{gcat_get_model}
\title{Gets a model.}
\usage{
gcat_get_model(projectId, locationId, modelDisplayName)
}
\arguments{
\item{projectId}{GCP project id}

\item{locationId}{location of GCP resources}

\item{modelDisplayName}{the name of the model shown in the interface}
}
\description{
Gets information about a trained model. Gets the most recent
model if multiple models are named the same
}
