% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{gcat_get_table_specs}
\alias{gcat_get_table_specs}
\title{Gets a table spec.}
\usage{
gcat_get_table_specs(projectId, locationId, displayName,
  fieldMask = NULL)
}
\arguments{
\item{projectId}{name of the dataset to get a table spec from}

\item{fieldMask}{Mask specifying which fields to read}
}
\description{
Gets a table spec.
}
