% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{gcat_global_dataset}
\alias{gcat_global_dataset}
\title{Set global dataset name}
\usage{
gcat_global_dataset(dataset)
}
\arguments{
\item{dataset}{dataset name you want this session to use by default or a
dataset object}
}
\value{
the dataset name (invisvibly)
}
\description{
set a dataset name used for this R session
}
\details{
This sets a dataset to a global environment value so you don't need to
supply the dataset argument to other API calls.
}
