% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{gcat_import_data}
\alias{gcat_import_data}
\title{Import data into AutoML Tables
https://cloud.google.com/automl-tables/docs/datasets#automl-tables-example-cli-curl}
\usage{
gcat_import_data(projectId, location, dataset_display_name,
  input_source = c("bq", "gcs"), input_url)
}
\arguments{
\item{location}{location of GCP resources}
}
\description{
Import data into AutoML Tables
https://cloud.google.com/automl-tables/docs/datasets#automl-tables-example-cli-curl
}
