% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{gcat_list_column_specs}
\alias{gcat_list_column_specs}
\title{Lists column specs in a table spec.}
\usage{
gcat_list_column_specs(projectId, locationId, displayName,
  pageSize = NULL, filter = NULL, fieldMask = NULL,
  pageToken = NULL)
}
\arguments{
\item{locationId}{location of GCP resources}

\item{pageSize}{Requested page size}

\item{filter}{Filter expression, see go/filtering}

\item{fieldMask}{Mask specifying which fields to read}

\item{pageToken}{A token identifying a page of results for the server to return}

\item{parent}{The resource name of the table spec to list column specs from}
}
\description{
Lists column specs in a table spec.
}
