% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{gcat_list_locations}
\alias{gcat_list_locations}
\title{Lists information about the supported locations for this service.}
\usage{
gcat_list_locations(projectId, pageToken = NULL, pageSize = NULL,
  filter = NULL)
}
\arguments{
\item{projectId}{the GCP project Id containing locations to fetch}

\item{pageToken}{The standard list page token}

\item{pageSize}{The standard list page size}

\item{filter}{The standard list filter}
}
\description{
Lists information about the supported locations for this service.
}
