% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{gcat_list_model_evaluations}
\alias{gcat_list_model_evaluations}
\title{Lists model evaluations.}
\usage{
gcat_list_model_evaluations(
  projectId = gcat_project_get(),
  locationId = gcat_region_get(),
  modelDisplayName
)
}
\arguments{
\item{projectId}{GCP project id}

\item{locationId}{the location Id}

\item{modelDisplayName}{the name of the model shown in the interface}
}
\description{
Lists model evaluations.
}
\seealso{
Other Model functions: 
\code{\link{gcat_batch_predict}()},
\code{\link{gcat_create_model}()},
\code{\link{gcat_get_model}()},
\code{\link{gcat_list_models}()}
}
\concept{Model functions}
