% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{gcat_list_models}
\alias{gcat_list_models}
\title{List Models}
\usage{
gcat_list_models(projectId = gcat_project_get(),
  locationId = gcat_region_get())
}
\arguments{
\item{projectId}{GCP project id}

\item{locationId}{location of GCP resources}
}
\description{
List Models
}
\seealso{
Other Model functions: \code{\link{gcat_create_model}},
  \code{\link{gcat_get_model}}
}
\concept{Model functions}
