% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{gcat_list_table_specs}
\alias{gcat_list_table_specs}
\title{Lists table specs in a dataset}
\usage{
gcat_list_table_specs(projectId, location, dataset_display_name)
}
\description{
Lists table specs in a dataset
}
