% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{gcat_location_path}
\alias{gcat_location_path}
\title{Create location path for use in other fucntions}
\usage{
gcat_location_path(projectId, location)
}
\arguments{
\item{projectId}{project containing datasets to list}

\item{location}{location of GCP resources}
}
\description{
Create location path for use in other fucntions
}
