% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{gcat_project_set}
\alias{gcat_project_set}
\title{Set the projectId for your AutoML Tables services}
\usage{
gcat_project_set(projectId)
}
\arguments{
\item{projectId}{The projectId}
}
\description{
Can also use environment argument GCAT_DEFAULT_PROJECT_ID
}
