% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{gcat_region_set}
\alias{gcat_region_set}
\title{Set the region}
\usage{
gcat_region_set(region = c("us-central1", "eu"))
}
\arguments{
\item{region}{Region for the endpoint}
}
\description{
Can also use environment argument GCAT_DEFAULT_REGION
}
