% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{GcrDestination}
\alias{GcrDestination}
\title{GcrDestination Object}
\usage{
GcrDestination(outputUri = NULL)
}
\arguments{
\item{outputUri}{Required}
}
\value{
GcrDestination object
}
\description{
GcrDestination Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The GCR location where the image must be pushed to.
}
\concept{GcrDestination functions}
