% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{GcsDestination}
\alias{GcsDestination}
\title{GcsDestination Object}
\usage{
GcsDestination(outputUriPrefix = NULL)
}
\arguments{
\item{outputUriPrefix}{Required}
}
\value{
GcsDestination object
}
\description{
GcsDestination Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The Google Cloud Storage location where the output is to be written to.
}
\concept{GcsDestination functions}
