% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{GcsSource}
\alias{GcsSource}
\title{GcsSource Object}
\usage{
GcsSource(inputUris = NULL)
}
\arguments{
\item{inputUris}{Required}
}
\value{
GcsSource object
}
\description{
GcsSource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The Google Cloud Storage location for the input content.
}
\concept{GcsSource functions}
