% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{GranularStats.buckets}
\alias{GranularStats.buckets}
\title{GranularStats.buckets Object}
\usage{
GranularStats.buckets()
}
\value{
GranularStats.buckets object
}
\description{
GranularStats.buckets Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A map from granularity key to example count for that key.E.g. for hour_of_day `13` means 1pm, or for month_of_year `5` means May).
}
\seealso{
Other GranularStats functions: \code{\link{GranularStats}}
}
\concept{GranularStats functions}
