% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{GranularStats}
\alias{GranularStats}
\title{GranularStats Object}
\usage{
GranularStats(GranularStats.buckets = NULL, buckets = NULL)
}
\arguments{
\item{GranularStats.buckets}{The \link{GranularStats.buckets} object or list of objects}

\item{buckets}{A map from granularity key to example count for that key}
}
\value{
GranularStats object
}
\description{
GranularStats Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Stats split by a defined in context granularity.
}
\seealso{
Other GranularStats functions: \code{\link{GranularStats.buckets}}
}
\concept{GranularStats functions}
