% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{Image}
\alias{Image}
\title{Image Object}
\usage{
Image(inputConfig = NULL, thumbnailUri = NULL, imageBytes = NULL)
}
\arguments{
\item{inputConfig}{An input config specifying the content of the image}

\item{thumbnailUri}{Output only}

\item{imageBytes}{Image content represented as a stream of bytes}
}
\value{
Image object
}
\description{
Image Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A representation of an image.Only images up to 30MB in size are supported.
}
\concept{Image functions}
