% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ImageClassificationDatasetMetadata}
\alias{ImageClassificationDatasetMetadata}
\title{ImageClassificationDatasetMetadata Object}
\usage{
ImageClassificationDatasetMetadata(classificationType = NULL)
}
\arguments{
\item{classificationType}{Required}
}
\value{
ImageClassificationDatasetMetadata object
}
\description{
ImageClassificationDatasetMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Dataset metadata that is specific to image classification.
}
\concept{ImageClassificationDatasetMetadata functions}
