% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ImageObjectDetectionAnnotation}
\alias{ImageObjectDetectionAnnotation}
\title{ImageObjectDetectionAnnotation Object}
\usage{
ImageObjectDetectionAnnotation(boundingBox = NULL, score = NULL)
}
\arguments{
\item{boundingBox}{Output only}

\item{score}{Output only}
}
\value{
ImageObjectDetectionAnnotation object
}
\description{
ImageObjectDetectionAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Annotation details for image object detection.
}
\concept{ImageObjectDetectionAnnotation functions}
