% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ImageObjectDetectionDatasetMetadata}
\alias{ImageObjectDetectionDatasetMetadata}
\title{ImageObjectDetectionDatasetMetadata Object}
\usage{
ImageObjectDetectionDatasetMetadata()
}
\value{
ImageObjectDetectionDatasetMetadata object
}
\description{
ImageObjectDetectionDatasetMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Dataset metadata specific to image object detection.
}
\concept{ImageObjectDetectionDatasetMetadata functions}
