% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ImageObjectDetectionEvaluationMetrics}
\alias{ImageObjectDetectionEvaluationMetrics}
\title{ImageObjectDetectionEvaluationMetrics Object}
\usage{

  ImageObjectDetectionEvaluationMetrics(boundingBoxMeanAveragePrecision = NULL,
  evaluatedBoundingBoxCount = NULL, boundingBoxMetricsEntries = NULL)
}
\arguments{
\item{boundingBoxMeanAveragePrecision}{Output only}

\item{evaluatedBoundingBoxCount}{Output only}

\item{boundingBoxMetricsEntries}{Output only}
}
\value{
ImageObjectDetectionEvaluationMetrics object
}
\description{
ImageObjectDetectionEvaluationMetrics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model evaluation metrics for image object detection problems.Evaluates prediction quality of labeled bounding boxes.
}
\concept{ImageObjectDetectionEvaluationMetrics functions}
