% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ImageObjectDetectionModelDeploymentMetadata}
\alias{ImageObjectDetectionModelDeploymentMetadata}
\title{ImageObjectDetectionModelDeploymentMetadata Object}
\usage{
ImageObjectDetectionModelDeploymentMetadata(nodeCount = NULL)
}
\arguments{
\item{nodeCount}{Input only}
}
\value{
ImageObjectDetectionModelDeploymentMetadata object
}
\description{
ImageObjectDetectionModelDeploymentMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model deployment metadata specific to Image Object Detection.
}
\concept{ImageObjectDetectionModelDeploymentMetadata functions}
