% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ImageObjectDetectionModelMetadata}
\alias{ImageObjectDetectionModelMetadata}
\title{ImageObjectDetectionModelMetadata Object}
\usage{
ImageObjectDetectionModelMetadata(nodeCount = NULL, nodeQps = NULL,
  modelType = NULL)
}
\arguments{
\item{nodeCount}{Output only}

\item{nodeQps}{Output only}

\item{modelType}{Optional}
}
\value{
ImageObjectDetectionModelMetadata object
}
\description{
ImageObjectDetectionModelMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model metadata specific to image object detection.
}
\concept{ImageObjectDetectionModelMetadata functions}
