% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ImportDataOperationMetadata}
\alias{ImportDataOperationMetadata}
\title{ImportDataOperationMetadata Object}
\usage{
ImportDataOperationMetadata()
}
\value{
ImportDataOperationMetadata object
}
\description{
ImportDataOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of ImportData operation.
}
\concept{ImportDataOperationMetadata functions}
