% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ImportDataRequest}
\alias{ImportDataRequest}
\title{ImportDataRequest Object}
\usage{
ImportDataRequest(inputConfig = NULL)
}
\arguments{
\item{inputConfig}{Required}
}
\value{
ImportDataRequest object
}
\description{
ImportDataRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for AutoMl.ImportData.
}
\seealso{
Other ImportDataRequest functions: \code{\link{projects.locations.datasets.importData}}
}
\concept{ImportDataRequest functions}
