% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{InputConfig.params}
\alias{InputConfig.params}
\title{InputConfig.params Object}
\usage{
InputConfig.params()
}
\value{
InputConfig.params object
}
\description{
InputConfig.params Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Additional domain-specific parameters describing the semantic of theimported data, any string must be up to 25000characters long.*  For Tables:   `schema_inference_version` - (integer) Required. The version of the       algorithm that should be used for the initial inference of the       schema (columns' DataTypes) of the table the data is being imported       into. Allowed values: '1'.
}
\seealso{
Other InputConfig functions: \code{\link{InputConfig}}
}
\concept{InputConfig functions}
