% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ListColumnSpecsResponse}
\alias{ListColumnSpecsResponse}
\title{ListColumnSpecsResponse Object}
\usage{
ListColumnSpecsResponse(columnSpecs = NULL, nextPageToken = NULL)
}
\arguments{
\item{columnSpecs}{The column specs read}

\item{nextPageToken}{A token to retrieve next page of results}
}
\value{
ListColumnSpecsResponse object
}
\description{
ListColumnSpecsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for AutoMl.ListColumnSpecs.
}
\concept{ListColumnSpecsResponse functions}
