% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ListDatasetsResponse}
\alias{ListDatasetsResponse}
\title{ListDatasetsResponse Object}
\usage{
ListDatasetsResponse(datasets = NULL, nextPageToken = NULL)
}
\arguments{
\item{datasets}{The datasets read}

\item{nextPageToken}{A token to retrieve next page of results}
}
\value{
ListDatasetsResponse object
}
\description{
ListDatasetsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for AutoMl.ListDatasets.
}
\concept{ListDatasetsResponse functions}
