% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ListModelEvaluationsResponse}
\alias{ListModelEvaluationsResponse}
\title{ListModelEvaluationsResponse Object}
\usage{
ListModelEvaluationsResponse(nextPageToken = NULL,
  modelEvaluation = NULL)
}
\arguments{
\item{nextPageToken}{A token to retrieve next page of results}

\item{modelEvaluation}{List of model evaluations in the requested page}
}
\value{
ListModelEvaluationsResponse object
}
\description{
ListModelEvaluationsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for AutoMl.ListModelEvaluations.
}
\concept{ListModelEvaluationsResponse functions}
