% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ModelEvaluation}
\alias{ModelEvaluation}
\title{Cloud AutoML API Objects 
Train high-quality custom machine learning models with minimum effort and machine learning expertise.}
\usage{
ModelEvaluation(displayName = NULL,
  classificationEvaluationMetrics = NULL, createTime = NULL,
  regressionEvaluationMetrics = NULL, name = NULL,
  evaluatedExampleCount = NULL, textSentimentEvaluationMetrics = NULL,
  imageObjectDetectionEvaluationMetrics = NULL,
  translationEvaluationMetrics = NULL, annotationSpecId = NULL,
  textExtractionEvaluationMetrics = NULL)
}
\arguments{
\item{displayName}{Output only}

\item{classificationEvaluationMetrics}{Model evaluation metrics for image, text, video and tables}

\item{createTime}{Output only}

\item{regressionEvaluationMetrics}{Model evaluation metrics for Tables regression}

\item{name}{Output only}

\item{evaluatedExampleCount}{Output only}

\item{textSentimentEvaluationMetrics}{Evaluation metrics for text sentiment models}

\item{imageObjectDetectionEvaluationMetrics}{Model evaluation metrics for image object detection}

\item{translationEvaluationMetrics}{Model evaluation metrics for translation}

\item{annotationSpecId}{Output only}

\item{textExtractionEvaluationMetrics}{Evaluation metrics for text extraction models}
}
\value{
ModelEvaluation object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2019-04-28 16:49:51
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
ModelEvaluation Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Evaluation results of a model.
}
\concept{ModelEvaluation functions}
