% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ModelExportOutputConfig.params}
\alias{ModelExportOutputConfig.params}
\title{ModelExportOutputConfig.params Object}
\usage{
ModelExportOutputConfig.params()
}
\value{
ModelExportOutputConfig.params object
}
\description{
ModelExportOutputConfig.params Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Additional model-type and format specific parameters describing therequirements for the to be exported model files, any string must be up to25000 characters long. * For `docker` format:    `cpu_architecture` - (string) 'x86_64' (default).    `gpu_architecture` - (string) 'none' (default), 'nvidia'.
}
\seealso{
Other ModelExportOutputConfig functions: \code{\link{ModelExportOutputConfig}}
}
\concept{ModelExportOutputConfig functions}
