% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{NormalizedVertex}
\alias{NormalizedVertex}
\title{NormalizedVertex Object}
\usage{
NormalizedVertex(x = NULL, y = NULL)
}
\arguments{
\item{x}{Required}

\item{y}{Required}
}
\value{
NormalizedVertex object
}
\description{
NormalizedVertex Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A vertex represents a 2D point in the image.The normalized vertex coordinates are between 0 to 1 fractions relative tothe original plane (image, video). E.g. if the plane (e.g. whole image) wouldhave size 10 x 20 then a point with normalized coordinates (0.1, 0.3) wouldbe at the position (1, 6) on that plane.
}
\concept{NormalizedVertex functions}
