% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{PredictRequest.params}
\alias{PredictRequest.params}
\title{PredictRequest.params Object}
\usage{
PredictRequest.params()
}
\value{
PredictRequest.params object
}
\description{
PredictRequest.params Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Additional domain-specific parameters, any string must be up to 25000characters long.*  For Image Classification:   `score_threshold` - (float) A value from 0.0 to 1.0. When the model    makes predictions for an image, it will only produce results that have    at least this confidence score. The default is 0.5.*  For Tables:   `feature_importance` - (boolean) Whetherfeature_importance       should be populated in the returnedTablesAnnotation(-s).       The default is false.
}
\seealso{
Other PredictRequest functions: \code{\link{PredictRequest}},
  \code{\link{projects.locations.models.predict}}
}
\concept{PredictRequest functions}
