% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{PredictRequest}
\alias{PredictRequest}
\title{PredictRequest Object}
\usage{
PredictRequest(PredictRequest.params = NULL, params = NULL,
  payload = NULL)
}
\arguments{
\item{PredictRequest.params}{The \link{PredictRequest.params} object or list of objects}

\item{params}{Additional domain-specific parameters, any string must be up to 25000}

\item{payload}{Required}
}
\value{
PredictRequest object
}
\description{
PredictRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for PredictionService.Predict.
}
\seealso{
Other PredictRequest functions: \code{\link{PredictRequest.params}},
  \code{\link{projects.locations.models.predict}}
}
\concept{PredictRequest functions}
