% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{PredictResponse.metadata}
\alias{PredictResponse.metadata}
\title{PredictResponse.metadata Object}
\usage{
PredictResponse.metadata()
}
\value{
PredictResponse.metadata object
}
\description{
PredictResponse.metadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Additional domain-specific prediction response metadata.
}
\seealso{
Other PredictResponse functions: \code{\link{PredictResponse}}
}
\concept{PredictResponse functions}
