% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_functions.R
\name{projects.locations.datasets.create}
\alias{projects.locations.datasets.create}
\title{Creates a dataset.}
\usage{
projects.locations.datasets.create(Dataset, parent)
}
\arguments{
\item{Dataset}{The \link{Dataset} object to pass to this method}

\item{parent}{The resource name of the project to create the dataset for}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/automl}{Google Documentation}

Other Dataset functions: \code{\link{Dataset}},
  \code{\link{projects.locations.datasets.patch}}
}
\concept{Dataset functions}
