% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_functions.R
\name{projects.locations.datasets.exportData}
\alias{projects.locations.datasets.exportData}
\title{Exports dataset's data to the provided output location.Returns an empty response in theresponse field when it completes.}
\usage{
projects.locations.datasets.exportData(ExportDataRequest, name)
}
\arguments{
\item{ExportDataRequest}{The \link{ExportDataRequest} object to pass to this method}

\item{name}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/automl}{Google Documentation}

Other ExportDataRequest functions: \code{\link{ExportDataRequest}}
}
\concept{ExportDataRequest functions}
