% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{projects.locations.datasets.list}
\alias{projects.locations.datasets.list}
\title{Lists datasets in a project.}
\usage{
projects.locations.datasets.list(parent, filter = NULL,
  pageToken = NULL, pageSize = NULL)
}
\arguments{
\item{parent}{The resource name of the project from which to list datasets}

\item{filter}{An expression for filtering the results of the request}

\item{pageToken}{A token identifying a page of results for the server to return}

\item{pageSize}{Requested page size}
}
\description{
Lists datasets in a project.
}
