% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_functions.R
\name{projects.locations.datasets.tableSpecs.columnSpecs.get}
\alias{projects.locations.datasets.tableSpecs.columnSpecs.get}
\title{Gets a column spec.}
\usage{
projects.locations.datasets.tableSpecs.columnSpecs.get(name,
  fieldMask = NULL)
}
\arguments{
\item{name}{The resource name of the column spec to retrieve}

\item{fieldMask}{Mask specifying which fields to read}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/automl}{Google Documentation}
}
