% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{projects.locations.datasets.tableSpecs.get}
\alias{projects.locations.datasets.tableSpecs.get}
\title{Gets a table spec.}
\usage{
projects.locations.datasets.tableSpecs.get(name, fieldMask = NULL)
}
\arguments{
\item{name}{The resource name of the table spec to retrieve}

\item{fieldMask}{Mask specifying which fields to read}
}
\description{
Gets a table spec.
}
