% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_functions.R
\name{projects.locations.datasets.tableSpecs.patch}
\alias{projects.locations.datasets.tableSpecs.patch}
\title{Updates a table spec.}
\usage{
projects.locations.datasets.tableSpecs.patch(TableSpec, name,
  updateMask = NULL)
}
\arguments{
\item{TableSpec}{The \link{TableSpec} object to pass to this method}

\item{name}{Output only}

\item{updateMask}{The update mask applies to the resource}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/automl}{Google Documentation}

Other TableSpec functions: \code{\link{TableSpec}}
}
\concept{TableSpec functions}
