% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_functions.R
\name{projects.locations.list}
\alias{projects.locations.list}
\title{Lists information about the supported locations for this service.}
\usage{
projects.locations.list(name, pageToken = NULL, pageSize = NULL,
  filter = NULL)
}
\arguments{
\item{name}{The resource that owns the locations collection, if applicable}

\item{pageToken}{The standard list page token}

\item{pageSize}{The standard list page size}

\item{filter}{The standard list filter}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/automl}{Google Documentation}
}
