% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_functions.R
\name{projects.locations.models.create}
\alias{projects.locations.models.create}
\title{Creates a model.Returns a Model in the responsefield when it completes.When you create a model, several model evaluations are created for it:a global evaluation, and one evaluation for each annotation spec.}
\usage{
projects.locations.models.create(Model, parent)
}
\arguments{
\item{Model}{The \link{Model} object to pass to this method}

\item{parent}{Resource name of the parent project where the model is being created}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/automl}{Google Documentation}

Other Model functions: \code{\link{Model}}
}
\concept{Model functions}
