% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_functions.R
\name{projects.locations.models.deploy}
\alias{projects.locations.models.deploy}
\title{Deploys a model. If a model is already deployed, deploying it with thesame parameters has no effect. Deploying with different parametrs(as e.g. changingnode_number) will reset the deployment state without pausing the model's availability.Only applicable for Text Classification, Image Object Detection and Tables;all other domains manage deployment automatically.Returns an empty response in theresponse field when it completes.}
\usage{
projects.locations.models.deploy(DeployModelRequest, name)
}
\arguments{
\item{DeployModelRequest}{The \link{DeployModelRequest} object to pass to this method}

\item{name}{Resource name of the model to deploy}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/automl}{Google Documentation}

Other DeployModelRequest functions: \code{\link{DeployModelRequest}}
}
\concept{DeployModelRequest functions}
