% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_functions.R
\name{projects.locations.models.export}
\alias{projects.locations.models.export}
\title{Exports a trained, 'export-able', model to a user specified Google CloudStorage location. A model is considered export-able if and only if it hasan export format defined for it inModelExportOutputConfig.Returns an empty response in theresponse field when it completes.}
\usage{
projects.locations.models.export(ExportModelRequest, name)
}
\arguments{
\item{ExportModelRequest}{The \link{ExportModelRequest} object to pass to this method}

\item{name}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/automl}{Google Documentation}

Other ExportModelRequest functions: \code{\link{ExportModelRequest}}
}
\concept{ExportModelRequest functions}
