% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_functions.R
\name{projects.locations.models.predict}
\alias{projects.locations.models.predict}
\title{Perform an online prediction. The prediction result will be directlyreturned in the response.Available for following ML problems, and their expected request payloads:* Image Classification - Image in .JPEG, .GIF or .PNG format, image_bytes                         up to 30MB.* Image Object Detection - Image in .JPEG, .GIF or .PNG format, image_bytes                           up to 30MB. * Text Classification - TextSnippet, content up to 60,000 characters,                        UTF-8 encoded.* Text Extraction - TextSnippet, content up to 30,000 characters,                    UTF-8 NFC encoded. * Translation - TextSnippet, content up to 25,000 characters, UTF-8                encoded.* Tables - Row, with column values matching the columns of the model,           up to 5MB. Not available for FORECASTINGprediction_type.}
\usage{
projects.locations.models.predict(PredictRequest, name)
}
\arguments{
\item{PredictRequest}{The \link{PredictRequest} object to pass to this method}

\item{name}{Name of the model requested to serve the prediction}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/automl}{Google Documentation}

Other PredictRequest functions: \code{\link{PredictRequest.params}},
  \code{\link{PredictRequest}}
}
\concept{PredictRequest functions}
