% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_functions.R
\name{projects.locations.operations.wait}
\alias{projects.locations.operations.wait}
\title{Waits for the specified long-running operation until it is done or reachesat most a specified timeout, returning the latest state.  If the operationis already done, the latest state is immediately returned.  If the timeoutspecified is greater than the default HTTP/RPC timeout, the HTTP/RPCtimeout is used.  If the server does not support this method, it returns`google.rpc.Code.UNIMPLEMENTED`.Note that this method is on a best-effort basis.  It may return the lateststate before the specified timeout (including immediately), meaning even animmediate response is no guarantee that the operation is done.}
\usage{
projects.locations.operations.wait(WaitOperationRequest, name)
}
\arguments{
\item{WaitOperationRequest}{The \link{WaitOperationRequest} object to pass to this method}

\item{name}{The name of the operation resource to wait on}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/automl}{Google Documentation}

Other WaitOperationRequest functions: \code{\link{WaitOperationRequest}}
}
\concept{WaitOperationRequest functions}
