% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{RegressionEvaluationMetrics}
\alias{RegressionEvaluationMetrics}
\title{RegressionEvaluationMetrics Object}
\usage{
RegressionEvaluationMetrics(rootMeanSquaredError = NULL,
  meanAbsoluteError = NULL, meanAbsolutePercentageError = NULL,
  rSquared = NULL)
}
\arguments{
\item{rootMeanSquaredError}{Output only}

\item{meanAbsoluteError}{Output only}

\item{meanAbsolutePercentageError}{Output only}

\item{rSquared}{Output only}
}
\value{
RegressionEvaluationMetrics object
}
\description{
RegressionEvaluationMetrics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metrics for regression problems.
}
\concept{RegressionEvaluationMetrics functions}
