% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{Row}
\alias{Row}
\title{Row Object}
\usage{
Row(values = NULL, columnSpecIds = NULL)
}
\arguments{
\item{values}{Required}

\item{columnSpecIds}{The resource IDs of the column specs describing the columns of the row}
}
\value{
Row object
}
\description{
Row Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A representation of a row in a relational table.
}
\concept{Row functions}
