% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{SingleCategoryStats}
\alias{SingleCategoryStats}
\title{SingleCategoryStats Object}
\usage{
SingleCategoryStats(count = NULL, value = NULL)
}
\arguments{
\item{count}{The number of occurrences of this value in the series}

\item{value}{The CATEGORY value}
}
\value{
SingleCategoryStats object
}
\description{
SingleCategoryStats Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The statistics of a single CATEGORY value.
}
\concept{SingleCategoryStats functions}
