% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{StringStats}
\alias{StringStats}
\title{StringStats Object}
\usage{
StringStats(topUnigramStats = NULL)
}
\arguments{
\item{topUnigramStats}{The statistics of the top 20 unigrams, ordered by}
}
\value{
StringStats object
}
\description{
StringStats Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The data statistics of a series of STRING values.
}
\concept{StringStats functions}
