% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{StructStats.fieldStats}
\alias{StructStats.fieldStats}
\title{StructStats.fieldStats Object}
\usage{
StructStats.fieldStats()
}
\value{
StructStats.fieldStats object
}
\description{
StructStats.fieldStats Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Map from a field name of the struct to data stats aggregated over seriesof all data in that field across all the structs.
}
\seealso{
Other StructStats functions: \code{\link{StructStats}}
}
\concept{StructStats functions}
