% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{StructStats}
\alias{StructStats}
\title{StructStats Object}
\usage{
StructStats(StructStats.fieldStats = NULL, fieldStats = NULL)
}
\arguments{
\item{StructStats.fieldStats}{The \link{StructStats.fieldStats} object or list of objects}

\item{fieldStats}{Map from a field name of the struct to data stats aggregated over series}
}
\value{
StructStats object
}
\description{
StructStats Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The data statistics of a series of STRUCT values.
}
\seealso{
Other StructStats functions: \code{\link{StructStats.fieldStats}}
}
\concept{StructStats functions}
