% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{StructType.fields}
\alias{StructType.fields}
\title{StructType.fields Object}
\usage{
StructType.fields()
}
\value{
StructType.fields object
}
\description{
StructType.fields Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Unordered map of struct field names to their data types.Fields cannot be added or removed via Update. Their names anddata types are still mutable.
}
\seealso{
Other StructType functions: \code{\link{StructType}}
}
\concept{StructType functions}
