% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{StructType}
\alias{StructType}
\title{StructType Object}
\usage{
StructType(StructType.fields = NULL, fields = NULL)
}
\arguments{
\item{StructType.fields}{The \link{StructType.fields} object or list of objects}

\item{fields}{Unordered map of struct field names to their data types}
}
\value{
StructType object
}
\description{
StructType Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
`StructType` defines the DataType-s of a STRUCT type.
}
\seealso{
Other StructType functions: \code{\link{StructType.fields}}
}
\concept{StructType functions}
