% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{TablesDatasetMetadata}
\alias{TablesDatasetMetadata}
\title{TablesDatasetMetadata Object}
\usage{

  TablesDatasetMetadata(TablesDatasetMetadata.targetColumnCorrelations = NULL,
  targetColumnSpecId = NULL, statsUpdateTime = NULL,
  targetColumnCorrelations = NULL, weightColumnSpecId = NULL,
  mlUseColumnSpecId = NULL, primaryTableSpecId = NULL)
}
\arguments{
\item{TablesDatasetMetadata.targetColumnCorrelations}{The \link{TablesDatasetMetadata.targetColumnCorrelations} object or list of objects}

\item{targetColumnSpecId}{column_spec_id of the primary table's column that should be used as the}

\item{statsUpdateTime}{The most recent timestamp when target_column_correlations field and all}

\item{targetColumnCorrelations}{Output only}

\item{weightColumnSpecId}{column_spec_id of the primary table's column that should be used as the}

\item{mlUseColumnSpecId}{column_spec_id of the primary table column which specifies a possible ML}

\item{primaryTableSpecId}{Output only}
}
\value{
TablesDatasetMetadata object
}
\description{
TablesDatasetMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata for a dataset used for AutoML Tables.
}
\seealso{
Other TablesDatasetMetadata functions: \code{\link{TablesDatasetMetadata.targetColumnCorrelations}}
}
\concept{TablesDatasetMetadata functions}
