% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{TablesModelColumnInfo}
\alias{TablesModelColumnInfo}
\title{TablesModelColumnInfo Object}
\usage{
TablesModelColumnInfo(columnSpecName = NULL, featureImportance = NULL,
  columnDisplayName = NULL)
}
\arguments{
\item{columnSpecName}{Output only}

\item{featureImportance}{Output only}

\item{columnDisplayName}{Output only}
}
\value{
TablesModelColumnInfo object
}
\description{
TablesModelColumnInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An information specific to given column and Tables Model, in contextof the Model and the predictions created by it.
}
\concept{TablesModelColumnInfo functions}
