% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{TablesModelMetadata}
\alias{TablesModelMetadata}
\title{TablesModelMetadata Object}
\usage{
TablesModelMetadata(trainBudgetMilliNodeHours = NULL,
  trainCostMilliNodeHours = NULL, targetColumnSpec = NULL,
  optimizationObjective = NULL, tablesModelColumnInfo = NULL,
  inputFeatureColumnSpecs = NULL)
}
\arguments{
\item{trainBudgetMilliNodeHours}{Required}

\item{trainCostMilliNodeHours}{Output only}

\item{targetColumnSpec}{Column spec of the dataset's primary table's column the model is}

\item{optimizationObjective}{Objective function the model is optimizing towards}

\item{tablesModelColumnInfo}{Output only}

\item{inputFeatureColumnSpecs}{Column specs of the dataset's primary table's columns, on which}
}
\value{
TablesModelMetadata object
}
\description{
TablesModelMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model metadata specific to AutoML Tables.
}
\concept{TablesModelMetadata functions}
