% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{TextClassificationDatasetMetadata}
\alias{TextClassificationDatasetMetadata}
\title{TextClassificationDatasetMetadata Object}
\usage{
TextClassificationDatasetMetadata(classificationType = NULL)
}
\arguments{
\item{classificationType}{Required}
}
\value{
TextClassificationDatasetMetadata object
}
\description{
TextClassificationDatasetMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Dataset metadata for classification.
}
\concept{TextClassificationDatasetMetadata functions}
