% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{TextClassificationModelMetadata}
\alias{TextClassificationModelMetadata}
\title{TextClassificationModelMetadata Object}
\usage{
TextClassificationModelMetadata()
}
\value{
TextClassificationModelMetadata object
}
\description{
TextClassificationModelMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model metadata that is specific to text classification.
}
\concept{TextClassificationModelMetadata functions}
