% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{TextExtractionAnnotation}
\alias{TextExtractionAnnotation}
\title{TextExtractionAnnotation Object}
\usage{
TextExtractionAnnotation(score = NULL, textSegment = NULL)
}
\arguments{
\item{score}{Output only}

\item{textSegment}{Required}
}
\value{
TextExtractionAnnotation object
}
\description{
TextExtractionAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Annotation for identifying spans of text.
}
\concept{TextExtractionAnnotation functions}
