% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{TextExtractionDatasetMetadata}
\alias{TextExtractionDatasetMetadata}
\title{TextExtractionDatasetMetadata Object}
\usage{
TextExtractionDatasetMetadata()
}
\value{
TextExtractionDatasetMetadata object
}
\description{
TextExtractionDatasetMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Dataset metadata that is specific to text extraction
}
\concept{TextExtractionDatasetMetadata functions}
