% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{TextExtractionEvaluationMetrics}
\alias{TextExtractionEvaluationMetrics}
\title{TextExtractionEvaluationMetrics Object}
\usage{
TextExtractionEvaluationMetrics(auPrc = NULL,
  confidenceMetricsEntries = NULL)
}
\arguments{
\item{auPrc}{Output only}

\item{confidenceMetricsEntries}{Output only}
}
\value{
TextExtractionEvaluationMetrics object
}
\description{
TextExtractionEvaluationMetrics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model evaluation metrics for text extraction problems.
}
\concept{TextExtractionEvaluationMetrics functions}
