% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{TextExtractionModelMetadata}
\alias{TextExtractionModelMetadata}
\title{TextExtractionModelMetadata Object}
\usage{
TextExtractionModelMetadata()
}
\value{
TextExtractionModelMetadata object
}
\description{
TextExtractionModelMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model metadata that is specific to text extraction.
}
\concept{TextExtractionModelMetadata functions}
