% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{TextSentimentAnnotation}
\alias{TextSentimentAnnotation}
\title{TextSentimentAnnotation Object}
\usage{
TextSentimentAnnotation(sentiment = NULL)
}
\arguments{
\item{sentiment}{Output only}
}
\value{
TextSentimentAnnotation object
}
\description{
TextSentimentAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains annotation details specific to text sentiment.
}
\concept{TextSentimentAnnotation functions}
