% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{TextSentimentDatasetMetadata}
\alias{TextSentimentDatasetMetadata}
\title{TextSentimentDatasetMetadata Object}
\usage{
TextSentimentDatasetMetadata(sentimentMax = NULL)
}
\arguments{
\item{sentimentMax}{Required}
}
\value{
TextSentimentDatasetMetadata object
}
\description{
TextSentimentDatasetMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Dataset metadata for text sentiment.
}
\concept{TextSentimentDatasetMetadata functions}
