% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{TextSentimentEvaluationMetrics}
\alias{TextSentimentEvaluationMetrics}
\title{TextSentimentEvaluationMetrics Object}
\usage{
TextSentimentEvaluationMetrics(confusionMatrix = NULL,
  quadraticKappa = NULL, meanAbsoluteError = NULL, recall = NULL,
  annotationSpecId = NULL, linearKappa = NULL, f1Score = NULL,
  precision = NULL, meanSquaredError = NULL)
}
\arguments{
\item{confusionMatrix}{Output only}

\item{quadraticKappa}{Output only}

\item{meanAbsoluteError}{Output only}

\item{recall}{Output only}

\item{annotationSpecId}{Output only}

\item{linearKappa}{Output only}

\item{f1Score}{Output only}

\item{precision}{Output only}

\item{meanSquaredError}{Output only}
}
\value{
TextSentimentEvaluationMetrics object
}
\description{
TextSentimentEvaluationMetrics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model evaluation metrics for text sentiment problems.
}
\concept{TextSentimentEvaluationMetrics functions}
