% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{TextSentimentModelMetadata}
\alias{TextSentimentModelMetadata}
\title{TextSentimentModelMetadata Object}
\usage{
TextSentimentModelMetadata()
}
\value{
TextSentimentModelMetadata object
}
\description{
TextSentimentModelMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model metadata that is specific to text sentiment.
}
\concept{TextSentimentModelMetadata functions}
