% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{TimestampStats}
\alias{TimestampStats}
\title{TimestampStats Object}
\usage{
TimestampStats(TimestampStats.granularStats = NULL,
  granularStats = NULL)
}
\arguments{
\item{TimestampStats.granularStats}{The \link{TimestampStats.granularStats} object or list of objects}

\item{granularStats}{The string key is the pre-defined granularity}
}
\value{
TimestampStats object
}
\description{
TimestampStats Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The data statistics of a series of TIMESTAMP values.
}
\seealso{
Other TimestampStats functions: \code{\link{TimestampStats.granularStats}}
}
\concept{TimestampStats functions}
