% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{TranslationAnnotation}
\alias{TranslationAnnotation}
\title{TranslationAnnotation Object}
\usage{
TranslationAnnotation(translatedContent = NULL)
}
\arguments{
\item{translatedContent}{Output only}
}
\value{
TranslationAnnotation object
}
\description{
TranslationAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Annotation details specific to translation.
}
\concept{TranslationAnnotation functions}
