% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{UndeployModelOperationMetadata}
\alias{UndeployModelOperationMetadata}
\title{UndeployModelOperationMetadata Object}
\usage{
UndeployModelOperationMetadata()
}
\value{
UndeployModelOperationMetadata object
}
\description{
UndeployModelOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of UndeployModel operation.
}
\concept{UndeployModelOperationMetadata functions}
