% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{UndeployModelRequest}
\alias{UndeployModelRequest}
\title{UndeployModelRequest Object}
\usage{
UndeployModelRequest()
}
\value{
UndeployModelRequest object
}
\description{
UndeployModelRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for AutoMl.UndeployModel.
}
\seealso{
Other UndeployModelRequest functions: \code{\link{projects.locations.models.undeploy}}
}
\concept{UndeployModelRequest functions}
