% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{UnigramStats}
\alias{UnigramStats}
\title{UnigramStats Object}
\usage{
UnigramStats(value = NULL, count = NULL)
}
\arguments{
\item{value}{The unigram}

\item{count}{The number of occurrences of this unigram in the series}
}
\value{
UnigramStats object
}
\description{
UnigramStats Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The statistics of a unigram.
}
\concept{UnigramStats functions}
