% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{VideoClassificationAnnotation}
\alias{VideoClassificationAnnotation}
\title{VideoClassificationAnnotation Object}
\usage{
VideoClassificationAnnotation(classificationAnnotation = NULL,
  type = NULL, timeSegment = NULL)
}
\arguments{
\item{classificationAnnotation}{Output only}

\item{type}{Output only}

\item{timeSegment}{Output only}
}
\value{
VideoClassificationAnnotation object
}
\description{
VideoClassificationAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains annotation details specific to video classification.
}
\concept{VideoClassificationAnnotation functions}
