% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{VideoClassificationDatasetMetadata}
\alias{VideoClassificationDatasetMetadata}
\title{VideoClassificationDatasetMetadata Object}
\usage{
VideoClassificationDatasetMetadata()
}
\value{
VideoClassificationDatasetMetadata object
}
\description{
VideoClassificationDatasetMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Dataset metadata specific to video classification.All Video Classification datasets are treated as multi label.
}
\concept{VideoClassificationDatasetMetadata functions}
