% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{VideoClassificationModelMetadata}
\alias{VideoClassificationModelMetadata}
\title{VideoClassificationModelMetadata Object}
\usage{
VideoClassificationModelMetadata()
}
\value{
VideoClassificationModelMetadata object
}
\description{
VideoClassificationModelMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model metadata specific to video classification.
}
\concept{VideoClassificationModelMetadata functions}
