% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{WaitOperationRequest}
\alias{WaitOperationRequest}
\title{WaitOperationRequest Object}
\usage{
WaitOperationRequest(timeout = NULL)
}
\arguments{
\item{timeout}{The maximum duration to wait before timing out}
}
\value{
WaitOperationRequest object
}
\description{
WaitOperationRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request message for Operations.WaitOperation.
}
\seealso{
Other WaitOperationRequest functions: \code{\link{projects.locations.operations.wait}}
}
\concept{WaitOperationRequest functions}
