% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{Dataset}
\alias{Dataset}
\title{Dataset Object}
\usage{
Dataset(createTime = NULL, textExtractionDatasetMetadata = NULL,
  videoClassificationDatasetMetadata = NULL,
  tablesDatasetMetadata = NULL, description = NULL,
  imageClassificationDatasetMetadata = NULL, etag = NULL,
  textSentimentDatasetMetadata = NULL, name = NULL,
  imageObjectDetectionDatasetMetadata = NULL,
  translationDatasetMetadata = NULL, exampleCount = NULL,
  videoObjectTrackingDatasetMetadata = NULL, displayName = NULL,
  textClassificationDatasetMetadata = NULL)
}
\arguments{
\item{createTime}{Output only}

\item{textExtractionDatasetMetadata}{Metadata for a dataset used for text extraction}

\item{videoClassificationDatasetMetadata}{Metadata for a dataset used for video classification}

\item{tablesDatasetMetadata}{Metadata for a dataset used for Tables}

\item{description}{User-provided description of the dataset}

\item{imageClassificationDatasetMetadata}{Metadata for a dataset used for image classification}

\item{etag}{Used to perform consistent read-modify-write updates}

\item{textSentimentDatasetMetadata}{Metadata for a dataset used for text sentiment}

\item{name}{Output only}

\item{imageObjectDetectionDatasetMetadata}{Metadata for a dataset used for image object detection}

\item{translationDatasetMetadata}{Metadata for a dataset used for translation}

\item{exampleCount}{Output only}

\item{videoObjectTrackingDatasetMetadata}{Metadata for a dataset used for video object tracking}

\item{displayName}{Required}

\item{textClassificationDatasetMetadata}{Metadata for a dataset used for text classification}
}
\value{
Dataset object
}
\description{
Dataset Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A workspace for solving a single, particular machine learning (ML) problem.A workspace contains examples that may be annotated.
}
\seealso{
Other Dataset functions: \code{\link{gcat_create_dataset_do_call}}
}
\concept{Dataset functions}
