% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{gcat_create_dataset}
\alias{gcat_create_dataset}
\title{Creates a dataset}
\usage{
gcat_create_dataset(projectId, location, displayName, parent)
}
\arguments{
\item{location}{location of GCP resources}

\item{parent}{}
}
\description{
Creates a dataset
}
