% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset.R
\name{gcat_create_dataset}
\alias{gcat_create_dataset}
\title{Creates a dataset.}
\usage{
gcat_create_dataset(Dataset, parent)
}
\arguments{
\item{Dataset}{The \link{Dataset} object to pass to this method}

\item{parent}{The resource name of the project to create the dataset for}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/automl}{Google Documentation}

Other Dataset functions: \code{\link{Dataset}}
}
\concept{Dataset functions}
