% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{gcat_create_dataset}
\alias{gcat_create_dataset}
\title{Creates a dataset (DOES NOT WORK)}
\usage{
gcat_create_dataset(
  projectId = gcat_project_get(),
  locationId = gcat_region_get(),
  displayName
)
}
\arguments{
\item{projectId}{GCP project id}

\item{locationId}{location of GCP resources}

\item{displayName}{the name of the dataset that is shown in the interface.
The name can be up to 32 characters long and can consist only of ASCII
Latin letters A-Z and a-z, underscores (_), and ASCII digits 0-9.}
}
\description{
Currently broken, see \href{https://github.com/justinjm/googleCloudAutoMLTablesR/issues/1}{issue} for more.
}
