% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{gcat_create_dataset}
\alias{gcat_create_dataset}
\title{Creates a dataset}
\usage{
gcat_create_dataset(projectId, locationId, displayName)
}
\arguments{
\item{projectId}{GCP project id}

\item{locationId}{location of GCP resources}

\item{displayName}{the name of the dataset that is shown in the interface.
The name can be up to 32 characters long and can consist only of ASCII
Latin letters A-Z and a-z, underscores (_), and ASCII digits 0-9.}
}
\description{
Creates a dataset
}
