% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{gcat_create_model}
\alias{gcat_create_model}
\title{Creates a model.Returns a Model in the responsefield when it completes.When you create a model, several model evaluations are created for it:a global evaluation, and one evaluation for each annotation spec.}
\usage{
gcat_create_model(projectId, locationId, datasetDisplayName,
  columnDisplayName, modelDisplayName, trainBudgetMilliNodeHours = NULL,
  optimizationObjective = NULL, targetColumnSpecName = NULL)
}
\arguments{
\item{datasetDisplayName}{the full name of your dataset.}

\item{columnDisplayName}{XXXXXXXXX}

\item{modelDisplayName}{the name of the model.}

\item{trainBudgetMilliNodeHours}{number of milli-node-hours for training. For example, 1000 = 1 hour.}

\item{optimizationObjective}{with the metric to optimize (optional). See https://cloud.google.com/automl-tables/docs/train#opt-obj}

\item{targetColumnSpecName}{with the full column name of your target column (optional).}
}
\description{
Creates a model.Returns a Model in the responsefield when it completes.When you create a model, several model evaluations are created for it:a global evaluation, and one evaluation for each annotation spec.
}
\concept{Model functions}
