% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{gcat_create_model}
\alias{gcat_create_model}
\title{Creates a model}
\usage{
gcat_create_model(
  projectId = gcat_project_get(),
  locationId = gcat_region_get(),
  datasetDisplayName = gcat_get_global_dataset(),
  columnDisplayName,
  modelDisplayName,
  trainBudgetMilliNodeHours = NULL,
  optimizationObjective = NULL,
  targetColumnSpecName = NULL
)
}
\arguments{
\item{projectId}{GCP project id}

\item{locationId}{location of GCP resources}

\item{datasetDisplayName}{the full name of your dataset}

\item{columnDisplayName}{the full name of the label or target}

\item{modelDisplayName}{the name of the model shown in the interface}

\item{trainBudgetMilliNodeHours}{number of milli-node-hours for training. For example, 1000 = 1 hour.}

\item{optimizationObjective}{with the metric to optimize (optional). See https://cloud.google.com/automl-tables/docs/train#opt-obj}

\item{targetColumnSpecName}{with the full column name of your target column (optional).}
}
\description{
Returns a Model in the responsefield when it completes. When you create a
model, several model evaluations are created for it:
a global evaluation, and one evaluation for each annotation spec.
}
\seealso{
Other Model functions: 
\code{\link{gcat_get_model}()},
\code{\link{gcat_list_model_evaluations}()},
\code{\link{gcat_list_models}()}
}
\concept{Model functions}
