% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{gcat_get_column_specs}
\alias{gcat_get_column_specs}
\title{Gets a column spec.}
\usage{
gcat_get_column_specs(projectId, locationId, displayName,
  columnDisplayName, fieldMask = NULL)
}
\arguments{
\item{locationId}{location of GCP resources}

\item{displayName}{}
}
\description{
Gets a column spec.
}
