% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{gcat_get_column_specs}
\alias{gcat_get_column_specs}
\title{Gets a column spec.}
\usage{
gcat_get_column_specs(projectId, locationId, displayName,
  columnDisplayName, fieldMask = NULL)
}
\arguments{
\item{projectId}{GCP project id}

\item{locationId}{location of GCP resources}

\item{displayName}{the name of the dataset that is shown in the interface.
The name can be up to 32 characters long and can consist only of ASCII
Latin letters A-Z and a-z, underscores (_), and ASCII digits 0-9.}

\item{columnDisplayName}{The name of the column to show in the interface.
The name can be up to 100 characters long and can consist only of ASCII
Latin letters A-Z and a-z, ASCII digits 0-9, underscores(_),
and forward slashes(/), and must start with a letter or a digit.}

\item{fieldMask}{Mask specifying which fields to read.}
}
\description{
Gets a column spec.
}
