% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{gcat_get_column_specs}
\alias{gcat_get_column_specs}
\title{Gets a column spec.}
\usage{
gcat_get_column_specs(projectId, location, datasetId, tableSpecId,
  fieldMask = NULL)
}
\description{
Gets a column spec.
}
