% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{gcat_get_dataset}
\alias{gcat_get_dataset}
\title{Gets a dataset}
\usage{
gcat_get_dataset(projectId, locationId, displayName)
}
\arguments{
\item{locationId}{location of GCP resources}
}
\description{
Gets a dataset
}
