% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{gcat_get_global_dataset}
\alias{gcat_get_global_dataset}
\title{Get global dataset name}
\usage{
gcat_get_global_dataset()
}
\value{
a string of the dataset displayName
}
\description{
Get global dataset name
}
\details{
Dataset name set this session to use by default. Set the dataset name
via \link{gcat_global_dataset}
}
