% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{gcat_get_location}
\alias{gcat_get_location}
\title{Get information about a location}
\usage{
gcat_get_location(projectId, locationId)
}
\arguments{
\item{projectId}{GCP project id}

\item{locationId}{location of GCP resources}
}
\description{
Get information about a location
}
\seealso{
Other loctions functions: \code{\link{gcat_list_locations}}
}
\concept{loctions functions}
