% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{gcat_list_datasets}
\alias{gcat_list_datasets}
\title{Check if objecti is Dataset object}
\usage{
gcat_list_datasets(projectId, location)
}
\arguments{
\item{location}{location of GCP resources}
}
\description{
Set global dataset name
}
\details{
set a dataset name used for this R session
Lists datasets in a project.
}
