% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{gcat_list_datasets}
\alias{gcat_list_datasets}
\title{Lists datasets in a project.}
\usage{
gcat_list_datasets(projectId = gcat_project_get(),
  locationId = gcat_region_get())
}
\arguments{
\item{projectId}{GCP project id}

\item{locationId}{location of GCP resources}
}
\description{
Lists datasets in a project.
}
