% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{gcat_list_locations}
\alias{gcat_list_locations}
\title{Lists information about the supported locations}
\usage{
gcat_list_locations(projectId)
}
\arguments{
\item{projectId}{GCP project Id to fetch a list of locations from}
}
\description{
Lists information about the supported locations
}
\seealso{
Other loctions functions: \code{\link{gcat_get_location}}
}
\concept{loctions functions}
