% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{gcat_list_table_specs}
\alias{gcat_list_table_specs}
\title{Lists table specs in a dataset}
\usage{
gcat_list_table_specs(projectId, locationId, displayName, filter = NULL,
  fieldMask = NULL, pageToken = NULL, pageSize = NULL)
}
\arguments{
\item{projectId}{name of the dataset to list table specs from}

\item{filter}{Filter expression, see go/filtering}

\item{fieldMask}{Mask specifying which fields to read}

\item{pageToken}{A token identifying a page of results for the server to return}

\item{pageSize}{Requested page size}
}
\description{
Lists table specs in a dataset
}
