% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{gcat_set_label}
\alias{gcat_set_label}
\title{Updates a dataset.}
\usage{
gcat_set_label(projectId, location, datasetId, tableSpecId,
  label_column_name, updateMask = NULL)
}
\arguments{
\item{location}{location of GCP resources}
}
\description{
Updates a dataset.
}
